requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
funcsigs==1.0.2
html5lib==0.9999999
Keras==2.0.6
Markdown==2.6.8
mock==2.0.0
numpy==1.13.1
pbr==3.1.1
protobuf>=3.5.0
PyYAML==3.12
scipy==0.19.1
six==1.10.0
tensorflow-gpu==1.4.0
Theano==0.9.0
Werkzeug==0.12.2
gensim==3.4.0
nltk==3.2.5
hunspell==0.5.4